+++ xhttp/xhttpclient/unmarshal.go
@@ -3,7 +3,9 @@ package xhttpclient
 import (
 	"net/http"
 
+	"github.com/xmidt-org/candlelight"
 	"github.com/xmidt-org/themis/config"
+	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 
 	"go.uber.org/fx"
 )
@@ -30,6 +32,9 @@ type ClientUnmarshalIn struct {
 	// RoundTripper is an optional http.RoundTripper component.  If present, this field will be used
 	// for clients unmarshalled by this instance.  Configuration will be ignored in favor of this component.
 	RoundTripper http.RoundTripper `optional:"true"`
+
+	// Tracing will be used to set up tracing instrumentation code.
+	Tracing candlelight.Tracing `optional:"true"`
 }
 
 // Unmarshal encompasses all the non-component information for unmarshalling and instantiating
@@ -69,6 +74,11 @@ func (u Unmarshal) Provide(in ClientUnmarshalIn) (Interface, error) {
 	} else {
 		rt = NewRoundTripper(o.Transport)
 	}
+
+	rt = otelhttp.NewTransport(rt,
+		otelhttp.WithPropagators(in.Tracing.Propagator()),
+		otelhttp.WithTracerProvider(in.Tracing.TracerProvider()),
+	)
 
 	chain := in.Chain.Extend(u.Chain)
 	if in.ChainFactory != nil {